+++ labgrid/remote/client.py
@@ -49,7 +49,7 @@ from ..util import diff_dict, flat_dict, dump, atomic_replace, labgrid_version, 
 from ..util.proxy import proxymanager
 from ..util.helper import processwrapper
 from ..driver import Mode, ExecutionError
-from ..logging import basicConfig, StepLogger
+from ..logging import basicConfig, CONSOLE, StepLogger
 
 # This is a workround for the gRPC issue
 # https://github.com/grpc/grpc/issues/38679.
@@ -2226,7 +2226,7 @@ def main():
     if args.verbose:
         logging.getLogger().setLevel(logging.INFO)
     if args.verbose > 1:
-        logging.getLogger().setLevel(logging.CONSOLE)
+        logging.getLogger().setLevel(CONSOLE)
     if args.debug or args.verbose > 2:
         logging.getLogger().setLevel(logging.DEBUG)